+++ lib/xrpl_flutter_sdk.dart
@@ -1,22 +1,24 @@
+// ignore_for_file: document_ignores, directives_ordering
+
 /// The first native Flutter/Dart SDK for the XRP Ledger (XRPL).
 ///
 /// Pure Dart, no platform channels.
 library;
 
-export 'src/codec/xrpl_base58.dart';
-export 'src/crypto/xrpl_ed25519.dart';
 // ------------------------------------------
 // Phase 1: Cryptographic fundamentals.
 // ------------------------------------------
 
 // Phase 1: Cryptographic fundamentals.
+export 'src/codec/xrpl_base58.dart';
 export 'src/crypto/xrpl_entropy.dart';
 export 'src/crypto/xrpl_hash.dart';
 export 'src/crypto/xrpl_key_algorithm.dart';
 export 'src/crypto/xrpl_secp256k1.dart';
 export 'src/crypto/xrpl_seed.dart';
+export 'src/crypto/xrpl_ed25519.dart';
 export 'src/exceptions/xrpl_crypto_exception.dart';
 export 'src/wallet/xrpl_wallet.dart';
 
 // Phase 2: Addresses.
-// export 'src/codec/xrpl_address_codec.dart';
+export 'src/address/xrpl_classic_address.dart';